requirements.txt
alembic==1.18.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==23.2.0
Babel==2.10.3
# bcc==0.29.1
blinker==1.7.0
Brlapi==0.8.5
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==25.1.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2vi
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
fastapi==0.128.0
greenlet==3.3.1
h11==0.16.0
httplib2==0.20.4
idna==3.6
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
meson==1.3.2
netaddr==0.8.0
netifaces==0.11.0
numpy==2.4.1
oauthlib==3.2.2
olefile==0.46
packaging==24.0
pexpect==4.9.0
pgvector==0.4.2
pillow==10.2.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pycairo==1.25.1
pycups==2.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.7.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu5.1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
setuptools==68.1.2
six==1.16.0
SQLAlchemy==2.0.46
SQLAlchemy-Utils==0.42.1
starlette==0.50.0
systemd-python==235
typing-inspection==0.4.2
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
uvicorn==0.40.0
wadllib==1.3.6
wheel==0.42.0
xdg==5
xkit==0.0.0
